requirements.txt
python-dotenv
pyrogram
tgcrypto
requests